pyproject.toml
[tool.poetry]
name = "bonus_calculator"
version = "0.1.0"
description = "Package to calculate a bonus, given a start date and end date."
authors = ["Thomas <tvanerven@intothesource.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyramid = "^2.0"
SQLAlchemy = "^1.4.27"
psycopg2-binary = "^2.9.0"
ipython = "^7.29.0"
factory-boy = "^3.2.1"
pytest = "^6.2.5"
waitress = "^2.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
